client/apicntext updates 2
